Proof explorer structural refactor
