+++ test-tools/proof-explorer/src/services/verification-service.tsx
@@ -1,0 +1,104 @@
+import { CredentialStatements, CredentialStatement, VerifiablePresentation } from '@concordium/web-sdk';
+import { VERIFIER_URL } from '../constants';
+import { Buffer } from 'buffer';
+import { WalletProvider } from './wallet-connection';
+import { useState } from 'react';
+import { TopLevelStatements } from '../types';
+
+export function getVerifierURL(): string {
+    return VERIFIER_URL;
+}
+
+async function submitProof(
+    statement: CredentialStatements,
+    provider: WalletProvider,
+    setMessages: (updateMessage: (oldMessages: string[]) => string[]) => void
+) {
+    let proof: VerifiablePresentation;
+    const challengeBuffer = new Uint8Array(32);
+    crypto.getRandomValues(challengeBuffer);
+    const challenge = Buffer.from(challengeBuffer).toString('hex');
+    console.log(statement);
+
+    try {
+        proof = await provider.requestVerifiablePresentation(challenge, statement);
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    } catch (err: any) {
+        if (err instanceof Error) {
+            setMessages((oldMessages) => [...oldMessages, `Could not get proof: ${err.message}`]);
+        } else {
+            console.log(err);
+        }
+        return;
+    }
+    console.log(proof.toString());
+    console.log(proof);
+    const resp = await fetch(`${getVerifierURL()}/v0/verify`, {
+        method: 'POST',
+        headers: {
+            'Content-Type': 'application/json',
+        },
+        body: proof.toString(),
+    });
+    if (resp.ok) {
+        setMessages((oldMessages) => [...oldMessages, 'Proof OK']);
+    } else {
+        const body = await resp.json();
+        setMessages((oldMessages) => [...oldMessages, `Proof not OK: (${resp.status}) ${body}`]);
+    }
+}
+
+export function SubmitProof(
+    all_statements: TopLevelStatements,
+    provider: WalletProvider | undefined
+): [(messages: string[]) => any, React.JSX.Element] {
+    const [messages, setMessages] = useState<string[]>([]);
+
+    const request = all_statements.map((s) => {
+        switch (s.type) {
+            case 'account':
+                return {
+                    statement: s.statement.statement,
+                    idQualifier: {
+                        type: 'cred',
+                        issuers: s.statement.idps.map((x) => x.id),
+                    },
+                } as CredentialStatement;
+            case 'web3id':
+                return {
+                    statement: s.statement.statement,
+                    idQualifier: {
+                        type: 'sci',
+                        issuers: s.statement.issuers,
+                    },
+                } as CredentialStatement;
+        }
+    });
+
+    return [
+        setMessages,
+        <div>
+            <div>
+                {provider !== undefined && (
+                    <button
+                        title="Submit the statement as a verified presentation request to the wallet."
+                        onClick={() => submitProof(request, provider, setMessages)}
+                        type="button"
+                        className="col-sm-4 btn btn-primary"
+                    >
+                        {'Prove'}
+                    </button>
+                )}
+            </div>
+            <hr />
+            <div>
+                <ol>
+                    {' '}
+                    {messages.map((m) => (
+                        <li className="alert alert-success"> {m} </li>
+                    ))}{' '}
+                </ol>
+            </div>
+        </div>,
+    ];
+}
